Refactor: controller: move some CIB-related functions
... from controld_execd.c to controld_cib.c, for better separation and
file size balance. No code changes, other than some functions are no longer
static.
